refactor(Contract/Edit):  edit sections now available with link
