+++ apps/web/src/app/[lang]/app/[type]/parties/[partyName]/[partyId]/contracts/[contractId]/contract/_components/contract-header.tsx
@@ -17,7 +17,7 @@ import { getMerchantContractHeaderMissingStepsByIdApi } from "src/app/[lang]/app
 import { deleteMerchantContractHeaderByIdApi } from "src/app/[lang]/app/actions/ContractService/delete-actions";
 import { postMerchantContractHeaderValidateByHeaderIdApi } from "src/app/[lang]/app/actions/ContractService/post-actions";
 import type { ContractServiceResource } from "src/language-data/ContractService";
-import ContractHeaderForm from "../../../contract-header-form";
+import ContractHeaderForm from "../../../_components/contract-header-form";
 
 export function ContractHeader({
   contractHeaderDetails,
