+++ apps/web/src/app/[lang]/app/[type]/parties/[partyName]/[partyId]/contracts/[contractId]/rebate-settings/_components/rebate-settings.tsx
@@ -36,7 +36,7 @@ import { handlePostResponse } from "src/app/[lang]/app/actions/api-utils-client"
 import {
   MerchantStoresWidget,
   RebateTableWidget,
-} from "../../../contract-widgets";
+} from "../../../_components/contract-widgets";
 
 export function RebateSettings({
   languageData,
