+++ apps/web/src/app/[lang]/app/[type]/parties/[partyName]/[partyId]/contracts/_components/contract-widgets.tsx
