+++ apps/web/src/app/[lang]/app/[type]/parties/[partyName]/[partyId]/contracts/_components/table-data.tsx
