+++ apps/web/src/app/[lang]/app/[type]/parties/[partyName]/[partyId]/contracts/new/page.tsx
@@ -8,7 +8,7 @@ import { getResourceData } from "src/language-data/ContractService";
 import { getBaseLink } from "src/utils";
 import { getRefundTableHeadersApi } from "src/app/[lang]/app/actions/ContractService/action";
 import PagePolicy from "src/app/[lang]/page-policy/page-policy";
-import ContractHeaderForm from "../contract-header-form";
+import ContractHeaderForm from "../_components/contract-header-form";
 
 export default async function Page({
   params,
