+++ apps/web/src/language-data/ContractService/resources/en.json
@@ -1,4 +1,9 @@
 {
+  "Contracts.Contract": "Contract",
+  "Contracts.RebateSettings": "Rebate settings",
+  "Contracts.Stores": "Stores",
+  "Contracts.ContractSettings": "Contract settings",
+
   "Refund": "Refund",
   "RefundTables": "Refund tables",
